Fix memory parsing
I pulled in the latest refinery code and it fails to parse this MaxAllow config value:

```yml
Collection:
  MaxAlloc: 1717600000 # 80% of 2GiB
```

The error is:

> Validation failed for config file refinery.yaml:
>   field Collection.MaxAlloc (1.7176e+06K) must be a valid memory size like '1Gb' or '100_000_000'

I tried updating to:
```yml
Collection:
  MaxAlloc: 1_717_600_000 # 80% of 2GiB
```

And that fails with:
> validateConfig unable to RELOAD config refinery.yaml: invalid size: 1_717_600_000

I think the regex was matching any character instead of periods, which is fixed here by escaping the period. It also was not matching underscores, which are now added to the match pattern.