feat: use a computed field for current descendant count in rules (#950)
## Which problem is this PR solving?

Metadata about a trace is added only to root spans. However, it's useful
to make sampling decisions before root span arrives. In this PR, I
addressed a specific use case described in [this issue](#646 )

## Short description of the changes

This PR adds a new concept called `ComputedField` in rule based sampler.
A `ComputedField` is a virtual field that's generated during rule
evaluation for a trace. The virtual field isn't actually defined on a
span.

After some discussion with Kent, we decided to use `?.` as the prefix
for computed field so that it shouldn't collide with any regular field
names.

This PR doesn't include documentation. I will address it in [this PR for
all rules related
documentation](https://github.com/honeycombio/refinery/pull/951/files).