maint: bump Husky (#966)
## Which problem is this PR solving?

Husky's [latest
release](https://github.com/honeycombio/husky/releases/tag/v0.24.0)
fixes two bugs:

* OTel Logs sent by JS over OTLP/JSON were having their ParentIDs
garbled because ... let's just call it an impedance mismatch between
signals. We now have special-case code to cope with it. This hasn't been
a problem because not many people send OTel logs, and not many of them
use JSON, and not many of them were sending from JS. We expect more of
that now that we're emphasizing frontend.
* If someone managed to construct an OTel signal with endtime before
starttime, we would end up interpreting that as a Very Large Amount Of
TIme. Now we won't.


## Short description of the changes

- bump Husky (and its dependencies)
- bump Go version
- update licenses because something went out of date