feat: Extract unique field names from sampler configs (#972)
To reduce the amount of across AZ zone data within a Refinery cluster,
we came up with an idea to only store information needed for making a
sampling decision in a central store. In order to extract those
information from a span, we need to have a list of fields used in
sampler configs.

This PR adds the code to generate the list mentioned above both during
Refinery startup and config reload by adding the generating code in the
`newFileConfig` function.

***Attention***
This PR is not merging into the current `main` branch. We are using a
feature branch for the dynamic scaling work