fix: make sure decision keep and drop is persisted in redis (#1020)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

Before this PR, decisions made for traces were not stored in redis. It
also wasn't part of the state machine when transition a trace from
awaiting for decision to either decision keep or drop.

## Short description of the changes

- Introduced DecisionDrop and DecisionKeep into the state machine
- check against in-memory cache before fetching from Redis for trace
status
- run smartwrapper tests against redis in the ci
- removed some unnecessary code