maint: Move Tracing config to config and do all the things (#1064)
To properly inject tracing, we needed the config, so this moves the
tracing config to the config area, sets up the metadata for it,
generates all the docs, and uses it. It also sets up injection of the
centralStore and the basicStore, making it possible to use them in
Refinery.