feat: implement span cache (#1068)
This contains two implementations of a span cache; `basic` is basically
just a map of traceID to trace. The second, `complex`, trades off
complexity and slightly slower performance for an implementation that
tries to avoid allocating memory most of the time. This may prove
valuable in a very high-throughput environment, but today I can't prove
it -- in all the benchmarks I've written, the *only* advantage is
low-allocation. So we have both implementations and we can play with
them in a working refinery.

One thing that I tried was to avoid the big sort in GetOldest; I
implemented an attempt to use a tdigest to find the threshold so that I
could avoid the sort, but the tdigest was more expensive than sorting,
even with 1 million items in the list, so I've abandoned that approach.
Sorting a slice of ints is really fast even for big lists.

I'm now using the same cache impact calculation as was already in use
for the previous span cache.

## Which problem is this PR solving?

The dynscaling system needs a new kind of cache for traces. This defines
the interface and also has two implementations to choose from.

## Short description of the changes

- Define an interface
- Basic implementation of span cache
- Complex implementation that avoids some memory allocations
- Tests and benchmarks