feat: initial implementation of CentralCollector (#1067)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

This is the skeleton code for integrating SmartStore into refinery for
dynamic scaling. There's quite bit of TODOs left for couple methods that
we still need to work out but I think this is a good starting point.

## Short description of the changes

- Created CentralCollector type
- Implemented AddSpan method
- Reworked the logic on how to extract key fields from each sampler