feat: make central collector injectable (#1078)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

The new CentralCollector should replace the old InMemoryCollector. 

## Short description of the changes

- Implements the rest of the collector functionalities except stress
relief
- Makes sure CentralCollector has access to all necessary information
when making a decision for a trace as well as
sending a kept trace.
- Since the `ParentID` is not currently used for anything but determine
whether a span is root span, I replaced the `ParentID` field with
`IsRoot` field on span objects
- Adapted applicable existing Refinery tests to use the new
CentralCollector.