feat: health checks now have alive/ready probes (#1100)
## Which problem is this PR solving?

Refinery previously only had a simple `alive` health check, that just
returned true all the time. In a world where we want to shut down
cleanly and drain our traffic, we need both `alive` and `ready` checks
so that we can turn off `ready` when we're shutting down. This way the
load balancer will stop sending traffic while the server drains itself.

## Short description of the changes

- Inject the new health system
- Call it from centralCollector in its different goroutines
- Call it with ready = false in the shutdown routine
- Rework alive endpoint to set the JSON content type
- Add the `/ready` endpoint (clone of alive) to deliver ready status
- Rewrite the grpc endpoints -- use the standard helper rather than
implementing the interface manually (as was done in #498)
- Delete the existing grpc tests because they relied on calling the
interface we no longer implement
- I haven't written new tests because I don't know how to do that in
grpc-land