feat: performance tweaks (#1110)
## Which problem is this PR solving?

- Addresses some performance issues with talking to Redis. Before these
changes, some calls to GetStatusForTraces were taking 800ms when called
with 5000 traceIDs. With this, 13000 traceIDs takes a little more than
100ms. I'm also seeing Refinery lasting quite a bit longer under heavy
loads, although it still has problems; I have more ideas but I'd like to
see this, plus #1108 and #1109 land first.
- This doesn't solve all performance issues, but it definitely helps.
See me for links to HC queries that show it.

## Short description of the changes

- GetStatusForTraces now takes a list of the statuses it cares about,
and only returns traces that have those values. This signature change
needed to be made in all the CentralStore interfaces, and implemented in
both redis and local stores.
- getTraceStatuses (which is called by GetStatusForTraces) now splits up
its redis requests into multiple goroutines that run in parallel, using
a fan-out/fan-in mechanism.
- Defaults for redis pool sizes have been increased to 15.
- Lots of tests have been fixed.