fix: prevent data race when retrieve information about cuckoo cache (#1115)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

In order to record metrics about dropped traces in cuckoo cache, we need
to access the cuckoofilter object periodically in a goroutine. Since
cuckoofilter itself is not concurrent safe, we need to add locks around
accessing it.

## Short description of the changes

Add mutex when accessing cuckoofilter for current count and load factor