fix: Update cache lookup to use read lock (#1145)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

- #1144 

## Short description of the changes

- Within the `get` method of an `environmentCache`, a read lock is added
around read access to the `items` map within the environment cache. It
is possible for one routine to hold a write lock within this function
while another routine enters the function and attempts to read from the
map while the other routine is writing to it.

---------

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>