feat: Add OTLP log endpoints (gRPC & HTTP) (#1187)
## Which problem is this PR solving?

Adds support for receiving OTLP logs. Only logs with a trace ID
associated are handled and are sampled along with spans. Logs without a
trace ID are forwarded to Honeycomb immediately like events without
TraceIDs are.

Further work to add support for configurable correlation IDs will be
done in later PRs.

**Note**: This is the 2.x version of adding OTLP Logs, see the following
PR that tracks adding to 3.x on main branch:
- #1168 

## Short description of the changes
- Add OTLP endpoints gRPC & HTTP (http/proto & http/json) including
support for gzip & zstd
- Adds missing API key check for trace gRPC handler with tests
- Add unit tests to verify endpoints and different encodings are handled
- Add unit tests to verify logs with a Trace ID are added to collector
and logs without a Trace ID are sent to libhoney for dispatch
- Rename processTraceRequest to `processOTLPRequest` and use in
otlp_trace
- Rename OTLP trace endpoint func postOTLP to postOTLPTraces
- Add MockCollector to aid with testing

---------

Co-authored-by: Pierre Tessier <pierre@pierretessier.com>
Co-authored-by: Kent Quirk <kentquirk@honeycomb.io>