feat: Working, tested, but unused pubsub system (#1205)
## Which problem is this PR solving?

- In order to implement a gossip protocol, we need a functional pubsub
system. I tried the go-redis library instead of redigo and it was pretty
simple to use.

## Short description of the changes

- Create a standardized pubsub interface
- Implement go-redis version of it
- Implement local version of it
- Write tests
- Write benchmarks
- Implement startstop
- Set up based on our redis peer config (I have not used all of the pool
size parameters, since they don't seem to matter anymore)

On my local machine (without going through an external network), it runs
at about 20K messages per second, with average latency of about 200uS
and max latency of about 5mS. On CI, the first test had avg latency of
6mS.

I also tried implementing a version using the reuidis library; see
comments in the code. Basically, for pubsub it was no faster and the
implementation was a bit ugly so I did not include it.