fix: allow sending otel tracing to non honeycomb backend (#1219)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?
`OtelTracing.APIKey` is specifically for setting a honeycomb api key and
header when setting up a tracer. When it's not set and
`OtelTracing.Enabled` is set true, Refinery should still set up a tracer
that can be used to send Refinery traces to non-honeycomb backend.

## Short description of the changes

- Only set APIKey headers when APIKey is not empty
- add some basic instrumentation