maint: change targeted arch to arm for local development Dockerfile (#1228)
<!--
Thank you for contributing to the project! 💜
Please make sure to:
- Chat with us first if this is a big change
  - Open a new issue (or comment on an existing one)
- We want to make sure you don't spend time implementing something we
might have to say No to
- Add unit tests
- Mention any relevant issues in the PR description (e.g. "Fixes #123")

Please see our [OSS process
document](https://github.com/honeycombio/home/blob/main/honeycomb-oss-lifecycle-and-practices.md#)
to get an idea of how we operate.
-->

## Which problem is this PR solving?

This dockerfile is used only for local development. Now most of us who
are actively working on this project are on ARM machines. Let's make it
easier to test refinery locally

## Short description of the changes

- Change targeted architecture to ARM