maint: Delete unused Dockerfile (#1292)
## Which problem is this PR solving?

- the repo contains a dockerfile that is not used in any of our CI or
`make` commands. It has been the source of confusion around our released
image's source. This Dockerfile was only intended to be used in local
`docker build .` commands, but has not outlived its usefulness.

## Short description of the changes

- Deletes the Dockerfile.


If alternatively, we want to keep the dockerfile, I could clean it up to
bring it in line with best security practices AND try to make it clearer
that it isn't used to build our images.