fix: use float histogram for otel metrics (#1303)
## Which problem is this PR solving?

- When we first introduced OTel metrics to Refinery, we only had int
histograms to work with. But now float ones are available and the int
one can yield very surprising results when your values are close to 0.

## Short description of the changes

- Replace Int64Histogram with Float64Histogram
- That's it.

We think that since we've never heard about this one before, it's
probably the case that most people using histograms are using them for
larger values, and so there won't be any real-life impact from this
change.