fix: Use peer transmission during redistribute and shutdown events (#1332)
## Which problem is this PR solving?

Updates the collector use the configured peer transmission to send spans
to peers during redistribute and shutdown events. This ensures metrics
for sending to Honeycomb and peers are accurate.

The peer transmission is already configured and used by the router when
receiving spans that should be sent to a peer. This is used the same
transmission object after a span was originally accepted but then should
be forwarded to a peer.

- Closes #1324

## Short description of the changes
- Inject the already configured peer transmission during collector init
- Update redistribute and shutdown logic to use the peer transmission
- Update tests throughout to verify the peer transmission is used
correctly and also verify it can be injected correctly
- note: there were existing tests to verify redistribute and shutdown
behaviour and have been updated