feat: Improve shutdown logic (#1347)
## Which problem is this PR solving?

- If part of Refinery has a problem shutting down, there's no way to
force it other than to do a `kill -9`, which is a pain.

## Short description of the changes

- Call the Stop() routines in the shutdown rather than in a defer()
- Instead of simply blocking on the signal channel, monitor that channel
in a goroutine. If more than one shutdown signal is received, exit
immediately without trying to be clean about it.
- Also, if a shutdown signal is received and the health system is
already "not alive", just die immediately; there's a good chance the
system won't be able to do an orderly shutdown.