fix: Allow the number of traces processed during each SendTicker tick to be configurable (#1393)
## Which problem is this PR solving?

- closes https://github.com/honeycombio/refinery/issues/1375

We found, by tracing the collect loop, that Refinery was sometimes doing
A TON of work each `SendTicker` tick depending on if there were a lot of
expired traces for which to make a sampling decision and send. Refinery
experienced _significant_ performance improvements by setting a max
value for how many expired traces would be processed each tick and by
moving the actual send into a separate goroutine.

## Short description of the changes

- `send` no longer actually sends all the spans in a trace. Instead, it
adds the read-to-send trace to a new channel. A new goroutine ranges
over this channel and does the actual work of preparing the spans and
adding them to the outgoing queue. This allows the collect loop to make
the sampling decision and get back to work faster, while the actually
send off can be handled independently.
- Adds a new configuration option, `Traces.MaxExpiredTraces` that limits
how many expired traces will be processed during a `SendTicker` tick.
The default value is 5000, which was the value we found kept the collect
loop consistently under 3 seconds.
- Fixes existing tests
- Adds a lot of tracing to the collect loop and its functions
- Adds new `collector_outgoing_queue` histogram to track the new queue's
size.

---------

Co-authored-by: Yingrong Zhao <22300958+VinozzZ@users.noreply.github.com>