maint: use a queue for events buffer in MockTransmission (#1401)
## Which problem is this PR solving?

Refinery's test suit has been flaky in the CI due to the fact that the
CI has way less CPU than our dev machines.

In the past, we have tried to use `Eventually` from the test framework
to stabilize our collector tests. Due to CPU constraint in the CI
environment and the amount of lock and unlock operations, tests have
been failing depends on when a goroutine is picked up by go scheduler.

This PR changes the event buffer in MockTransmission from a mutex
protected slice to a channel. This separates the `Enqueue` operation
from the `Get` operation so that a writer will not be blocked by a
reader

## Short description of the changes

- implement event buffer using a go channel
- update all existing tests to use the new semantic
- check whether we are in shutdown before enqueue a event from the
collector to libhoney