feat: Add a way to specify the team key for config fetches (experimental) (#1410)
## Which problem is this PR solving?

To fetch a configuration from Honeycomb, we're going to need a way to
specify a header for the team key. This allows the key to be specified
in an environment variable called `HONEYCOMB_CONFIG_KEY`. This is
deliberately different from the telemetry key because the key will have
to have special permissions.

If the key is not blank, then the header `x-honeycomb-team` will be set
and used for the config requests. The header will not be used if the key
is blank.

We are deliberately not adding this to public documentation because this
feature is still experimental and we may well change it.

## Short description of the changes

- Fetch `HONEYCOMB_CONFIG_KEY` from the environment
- Set `x-honeycomb-team` header if the key is non-blank.