fix: only load config and rules once (#1470)
## Which problem is this PR solving?
 
We found an issue where refinery, as part of its config load process on
startup or reload, would load the config/rules (whether from file or
http or any), multiple times. This was happening because we the config
data was passed around as a reader and if we ever needed the data a
second time (which we do for validation vs unmarshaling into config) we
had to go get another reader.

## Short description of the changes

- updates `newFileConfig` to first grab the data of all config locations
and the use this data for all validation and unmarshalling. This
prevents `newFileConfig` from needing to grab the data multiple times
- updated some unit tests, but most needed unchanged.

---------

Co-authored-by: Yingrong Zhao <22300958+VinozzZ@users.noreply.github.com>