test: improve existing integration benchmark, add new test and benchmark for Router.batch() (#1580)
## Which problem is this PR solving?
We need benchmarks which roughly approximate production bottlenecks in
order to evaluate the effectiveness of optimizations.

## Short description of the changes
Improves the pre-existing integration benchmarks to be more internally
efficient, and use msgpack+zstd which is more typical today. Note these
still output json which isn't typical in production so that still needs
to be fixed.

Also adds a new test and benchmark for Router.batch() which is the
primary consumer of both CPU and memory, at least in our internal
cluster.