perf: add lazy-loading MessagePack payload for events (#1586)
## Which problem is this PR solving?
Refinery is not efficient! In production profiles we often spend over
50% of CPU time allocating and freeing memory, and a good chunk of the
rest deserializing messagepack data.

## Short description of the changes
This is the first (and hopefully largest) change on the way to cutting
down the amount of time refinery spends deserializing events. In place
of the ubiquitous `map[string]any`, this keeps the original serialized
data with the event throughout its lifetime, and interrogates it when
necessary to retrieve individual fields. The theory here is that
ultimately refinery needs to access only a small fraction of fields, and
unpacking the rest into distinct scalars on the heap is unnecessary.
This is an approach which has served us well with other services.

Added here are the initial container types needed to make this happen,
and all the boilerplate interface changes that go with that.
Unfortunately I had to bring in a new message pack library, because our
old one doesn't know how to work with flat `[]byte` slices efficiently,
but it's also not a drop-in replacement. (The old library also hasn't
had a commit in two years and is looking like it may be abandoned.)

But, this PR itself does not improve performance. To see initial gains,
we'll need to optimize how we do messagepack deserialization, and
pre-request some of our fields instead of demanding them all one-by-one
with Get(). And this PR is big enough as it is. Longer term we'll also
want to leverage this pre-serialized data during re-transmission to
peers, and handle otel data efficiently in some way (probably direct
transcoding). For the record I think it would be better if we were using
otel for refinery's outbound transmission, but that's too big a change
if we want to get this done as well.

This is only one behavioral change here, which is that the event size
calculation now uses the serialized data when available, which will
definitely produce somewhat different numbers than what we do today.
Since this didn't break any tests I'm assuming that ok.

---------

Co-authored-by: Mike Goldsmith <goldsmith.mike@gmail.com>