feat: draft DirectTransmit to replace libhoney for trace data transmission (#1595)
## Which problem is this PR solving?
Libhoney is old and tries to be safe and have good ergonomics and
support a lot of options, and that means it also comes with a lot of
overhead. It's also not easily extensible, should we want to improve
refinery peer-to-peer communication.

## Short description of the changes
This is a first iteration of a wholesale replacement for libhoney (for
refinery's use only). It's not nearly as much code as libhoney, because
it doesn't do nearly as much stuff, but doing these sends directly make
some jobs - like handling send responses - _much_ simpler. It's not
ready to go yet though, note the TODOs here. I think this PR is big
enough already, and would like to land this and extend it in further
PRs. Note I'm not actually using this new code yet; the goal is to get
to a correct implementation (which this isn't because it doesn't honor
max event or batch byte sizes), then improve the performance and
scalability of the implementation.

This PR also extends the integration tests with a test API server. In
the past we didn't really need that since we leaned on libhoney's tested
transmission library, but rolling our own means we need our own testing.

I also removed the Flush() method from Transmission, because AFAICT it
doesn't do anything that we weren't doing already in Stop().