perf: use a buffer pool for http body reads (#1610)
## Which problem is this PR solving?
It sure is expensive allocating these big buffers for every HTTP read,
especially given that their size is initially unknown (at least when
compressed.) We also copied compressed bodies twice, which was totally
unnecessary.

We also did not apply any body size limit to uncompressed payloads.

## Short description of the changes
The main change here is to keep a pool of buffers for the HTTP bodies,
allowing us to easily avoid potentially large allocations. It also
restructures the read so that we only read the body contents at most
once, and applies the size limit to uncompressed payloads.

For our optimized Payload path, this does mean we need to make new
copies of the relevant sections of the body for storage with the
Payload. However, this is better than what we did before, because it's
less memory, the size is known a priori, and it solves the problem of a
single Event hanging onto a reference to the entire buffer from a much
larger batch.

We could take this further and also recycle Payload's messagepack
buffer, but that requires accurately identifying each and every place
where an Event goes out of scope - a job I'd rather leave to the garbage
collector, at least for now.