maint: relocate pure function config.IsLegacyAPIKey() to the file defining the Config interface (#1630)
## Which problem is this PR solving?

- I was confused briefly about the location of config.IsLegacyAPIKey(). 

## Short description of the changes

- Since it does not require anything from a real implementation of the
Config interface, move the pure function to the file that defines the
interface.
- Avoid implementation drift: add an assertion to the test cases to
confirm Refinery's legacy key detection behavior matches the key
detection in the husky library.

It's my hope that we can make the same performance improvement to the
key detection in husky and remove this implementation from Refinery.
Until then, make sure the two libraries agree on what's a legacy key.