test: improve efficiency of BenchmarkCollectorWithSamplers (#1650)
## Which problem is this PR solving?
This benchmark was totally bogged down in creating test data points,
making it slow to run and innaccurate.

## Short description of the changes
Changes the structure of the benchmark so that N is the number of spans,
not the number of traces. Gives up on the 1M-span trace which I think is
not really necessary to get the idea of the performance curve. Re-uses
spans.

At some point the input will again become the bottleneck here, but
multiple input goroutines really slowed things down so I stuck with just
the one - for now anyway.