perf: CollectLoop-local samplers, with shared dynsamplers (#1673)
Samplers include some expensive operations (mainly key building) which
are currently happening inside a mutex, defeating many of the potential
gains for concurrent collect loops.

Gives each collect loop a local set of samplers, as well as a reload
signal to rebuild them on config change. The dynsamplers, which can
contain global state related to total throughput, are still shared, but
this is now managed by the sampler factory. Each sampler has a function
defining how to create the dynsampler for that type, which is used by
the sampler factory. Unfortunately the penchant for lots of concrete
types and struct fields here makes this code a little more verbose than
it would otherwise need to be. On the plus side, the samplers themselves
get a little simpler, and the rules-based sampler now defers to the
sampler factory to make its sub-samplers, which I think it probably the
right long-term direction.

Includes a pretty ferocious new integration test for config reloads in
the app package. This revealed race conditions in the config code which
are also fixed here.