feat: config option deprecation (#1667)
## Which problem is this PR solving?

fix: #1642 
In the issue, the proposed solution is to return error if the
`latestversion` defined for a config option is older than the current
version. In this PR, we will hold off on that change since this is the
first time we introduce the concept of deprecation in Refinery. All
deprecated configs will log warnings for now.


## Short description of the changes

- Add a warning log for config fields that have `lastversion` value set
and the current Refinery version is before `lastversion`
- Log a warning for config fields that have `latestversion` value set
and the current Refinery version is after `latestversion`
- Suuport custom deprecation message through `deprectiontext` field in
`configMeta.yaml`


## Example output:
```
ERROR: Validation failed for config [tools/convert/minimal_config.yaml]:
  field General.DatasetPrefix (%%) must be purely alphanumeric
WARNING: Configuration warnings for [tools/convert/minimal_config.yaml]:
  field BufferSizes.PeerBufferSize is deprecated since version v2.9.7
  CacheCapacity is deprecated since version v2.9.7. Set PeerQueueSize and IncomingQueueSize instead.
  field BufferSizes.UpstreamBufferSize is deprecated since version v2.9.7
  Please update your configuration following the latest documentation here: https://docs.honeycomb.io/manage-data-volume/sample/honeycomb-refinery/configure/
```

---------

Co-authored-by: Kent Quirk <kentquirk@honeycomb.io>