feat(convert): support updating deprecated field using convert tool (#1674)
## Which problem is this PR solving?

This PR adds a new feature to the `convert` tool so that users can
easily updating their existing config when there's deprecated fields in
them.

## Short description of the changes

A new `--remove-deprecated` flag is added to support removing/replacing
deprecated fields. Deprecated fields are marked with`lastversion` in
configMeta.yaml. Each deprecated field can define `replacements` with
`formula` to calculate new values. The `convert` tool uses this metadata
to know what to remove and what to replace it with.

## Example output
```
## INPUT
General:
  ConfigurationVersion: 2

Collection:
  CacheCapacity: 5000       # Deprecated
  PeerQueueSize: 25000      # Not deprecated
  MaxAlloc: 2147483648      # Not deprecated

## OUTPUT
# The following deprecated config options were removed:
# - Collection.CacheCapacity (deprecated in v2.9.7)
# -   → IncomingQueueSize set to 15000 (using formula: 3 * value)
#
Collection:
    IncomingQueueSize: 15000
    MaxAlloc: 2147483648
    PeerQueueSize: 25000
General:
    ConfigurationVersion: 2

```