fix: dataset name should only be url encoded once (#1681)
## Which problem is this PR solving?

Customer has reported that their `service.name` with a space in the
value is showing up as `foo%20bar`

## Short description of the changes

Turns out there was a
[fix](https://github.com/honeycombio/libhoney-go/commit/9f4f37c3759e781e74a3c901417e80b50796a63e)
added for this issue in libhoney. Our new `DirectTransmit` didn't
include this fix. This PR introduce the fix into `DirectTransmit`.
TLDR is `url.Path` is expected to be unescaped. `url.String()` does the
escape by default. If you set url.Path to be foo%20bar, url.String()
will return foo%2520bar