fix: set default sizes for receive queues (#1694)
## Which problem is this PR solving?

Starting up Refinery 3.0-beta without IncomingQueueSize or
PeerQueueSize configured results in queue sizes set to `0`.

## Short description of the changes

The sizes for these queues used to be computed from the CacheCapacity
config option that has been removed.[1] With that math gone, we need to
declare defaults or these will be size 0 and then you're going to have a
bad time.

[1](https://github.com/honeycombio/refinery/pull/1666)